renames layout-store to overlay-store
